Implement PostgreSQL database connection hosted on Railway
This involved adding a controller, db pool and queries, script to populate the database, and custom form validation.

One tricky aspect of this expansion was figuring out that you had to double the single quotes for them to work with SQL, or else it would think it was the end of the string and give a syntax error.
This issue was addressed with a custom sanitizer which would replace one ' with ''.

Connecting to the database was relatively straightforward, and I used the public database URL as the environment variable.